building

+++ workspaces/assets/utils/compile-assets.sh
@@ -1,3 +1,8 @@
+if [ -f "./out/assets.rbxlx.lock" ]; then
+    echo "[ERROR] The asset file is currently being edited in Studio. Please close the asset.rbxlx file before compiling new assets."
+    exit 1
+fi
+
 rm -rf ./out
 mkdir out
